fix file-formatters parameter type, expand tests to include dump.to_zip, updated README
