+++ README.md
@@ -1018,6 +1018,10 @@ _Parameters_:
 - `pretty-descriptor`: Specifies how datapackage descriptor (`datapackage.json`) file will look like:
     - `False` (default) - descriptor will be written in one line.
     - `True` - descriptor will have indents and new lines for each key, so it becomes more human-readable.
+- `file-formatters`: Specifies custom file format handlers. An object with mapping of format name to Python module and class name.
+    - Allows to override the existing `csv` and `json` format handlers or add support for new formats.
+    - Note that such changes may make the resulting datapackage incompatible with the frictionlessdata specs and may cause interoperability problems.
+    - Example usage: [pipeline-spec.yaml](tests/cli/pipeline-spec.yaml) (under the `custom-formatters` pipeline), [XLSXFormat class](tests/cli/custom_formatters/xlsx_format.py)
 
 ### ***`dump.to_zip`***
 
@@ -1031,6 +1035,7 @@ _Parameters_:
 - `add-filehash-to-path` - Same as in `dump.to_path`
 - `counters` - Same as in `dump.to_path`
 - `pretty-descriptor` - Same as in `dump.to_path`
+- `file-formatters` - Same as in `dump.to_path`
 
 #### *Note*
